+++ R/SingleRfit.R
@@ -25,19 +25,6 @@ estimate_popsize.fit <- function(y,
   tbgname <- colnames(X)
   X <- as.matrix(X)
 
-  log_like <- family$make_minusloglike(y = y, X = X,
-                                       weight = prior.weights)
-  grad <- family$make_gradient(y = y, X = X,
-                               weight = prior.weights)
-  hessian <- family$make_hessian(y = y, X = X,
-                                 weight = prior.weights)
-
-  df.reduced <- length(y) - dim(X)[2]
-
-  if (family$family %in% c("zotnegbin", "ztnegbin")) {
-    df.reduced <- 2 * df.reduced
-  }
-
   if (method == "robust") {
 
     if (!is.null(dispersion)) {
@@ -54,26 +41,40 @@ estimate_popsize.fit <- function(y,
                                        start = start,
                                        silent = control$silent,
                                        disp.given = control$disp.given,
-                                       trace = control$trace)
+                                       trace = control$verbose)
 
     iter <- FITT$iter
     dispersion <- FITT$disp
     weights <- FITT$weights
     beta <- c(dispersion, FITT$coefficients)
   } else if (method == "mle") {
+    log_like <- family$make_minusloglike(y = y, X = X,
+                                         weight = prior.weights)
+    grad <- family$make_gradient(y = y, X = X,
+                                 weight = prior.weights)
+    
     weights <- prior.weights
     methodopt <- control$mleMethod
-    
+
     if (!isFALSE(control$optimPass)) {
       ctrl <- control$optimPass
     } else {
-      ctrl <- list(
-        factr = control$epsilon,
-        maxit = control$maxiter,
-        trace = if (is.numeric(control$trace)) control$trace else 0
-      )
+      if (methodopt == "L-BFGS-B") {
+        ctrl <- list(
+          factr = control$epsilon,
+          maxit = control$maxiter,
+          trace = control$verbose,
+          trace = if (is.numeric(control$trace)) control$trace else 0
+        )
+      } else {
+        ctrl <- list(
+          reltol = control$epsilon,
+          maxit = control$maxiter,
+          trace = control$verbose
+        )
+      }
     }
-    
+
     FITT <- stats::optim(fn = log_like,
                          par = start,
                          gr = function(x) -grad(x),
@@ -81,6 +82,9 @@ estimate_popsize.fit <- function(y,
                          control = ctrl)
     beta <- FITT$par
     iter <- FITT$counts
+    if (FITT$convergence == 1) {
+      warning("Convergence not obtained in ", control$maxiter, " iterations of mle fitting algorithm", sep = "")
+    }
   } else {
     stop("Method not implemented")
   }
@@ -99,14 +103,9 @@ estimate_popsize.fit <- function(y,
     names(beta) <- c("log(dispersion)", tbgname)
   }
 
-  hess <- hessian(beta)
-
-  list(ll = log_like,
-       grad = grad,
-       hessian = hessian,
-       beta = beta,
-       weights = weights,
-       hess = hess,
-       iter = iter,
-       degf = df.reduced)
+  list(
+    beta = beta,
+    weights = weights,
+    iter = iter
+  )
 }