+++ R/marginal.R
@@ -14,7 +14,7 @@ marginalFreq <- function(object,
                          onecount = NULL,
                          range) {
   if (missing(range)) {range <- (min(object$y):max(object$y))}
-  y <- table(object$y)[range]
+  y <- table(object$y)[names(table(object$y)) %in% as.character(range)]
   y <- y[!is.na(y)]
   trcount <- object$trcount
   if(!is.null(onecount)) {
@@ -52,7 +52,8 @@ marginalFreq <- function(object,
     names(res)[1] <- "0"
   }
   res <- structure(list(table = res, y = y, 
-                        df = length(y) - length(object$coefficients)), 
+                        df = length(y) - length(object$coefficients),
+                        name = object$model$family), 
                    class = c("singleRmargin"))
   res
 }