+++ NEWS.md
@@ -5,6 +5,10 @@
 * features:
   * control parameters for model
   * control parameters for regression in bootstrap sampling
+  * leave one out diagnostics for popsize and regression parameters (dfbetas were corrected)
+  * fixes for Goodness of fit tests in zero one truncated models
+  * computational improvements in IRLS
+  * other small bugfixes
 
 # singleRcapture 0.1.1